Moving cli, config and Main into subpackage app
* Cleaned some comments in Config
