+++ src/main/java/org/schemaspy/app/cli/CommandLineArguments.java
@@ -1,14 +1,15 @@
-package org.schemaspy.cli;
+package org.schemaspy.app.cli;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import org.schemaspy.app.config.SchemaSpyConfiguration;
 
 import java.io.File;
 
 /**
  * Holds all supported command line arguments.
  * <p>
- * An instance of this class registered as singleton bean in {@link org.schemaspy.SchemaSpyConfiguration} via {@link CommandLineArgumentParser}.
+ * An instance of this class registered as singleton bean in {@link SchemaSpyConfiguration} via {@link CommandLineArgumentParser}.
  * Therefore it can be injected in other beans, for example:
  * <p>
  * <pre>
