+++ src/main/java/org/schemaspy/app/cli/PropertyFileDefaultProvider.java
@@ -1,4 +1,4 @@
-package org.schemaspy.cli;
+package org.schemaspy.app.cli;
 
 import com.beust.jcommander.IDefaultProvider;
 import org.slf4j.Logger;
