+++ src/main/java/org/schemaspy/app/cli/PropertyFileDefaultProviderFactory.java
@@ -1,4 +1,4 @@
-package org.schemaspy.cli;
+package org.schemaspy.app.cli;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
