+++ src/main/java/org/schemaspy/service/TableService.java
@@ -1,7 +1,7 @@
 package org.schemaspy.service;
 
 import org.schemaspy.Config;
-import org.schemaspy.cli.CommandLineArguments;
+import org.schemaspy.app.cli.CommandLineArguments;
 import org.schemaspy.model.*;
 import org.schemaspy.model.xml.ForeignKeyMeta;
 import org.schemaspy.model.xml.TableColumnMeta;
