+++ src/test/java/org/schemaspy/app/cli/CommandLineArgumentParserTest.java
@@ -1,4 +1,4 @@
-package org.schemaspy.cli;
+package org.schemaspy.app.cli;
 
 import com.beust.jcommander.ParameterException;
 import org.junit.Test;
