+++ src/test/java/org/schemaspy/app/cli/ConfigFileArgumentParserTest.java
@@ -1,4 +1,4 @@
-package org.schemaspy.cli;
+package org.schemaspy.app.cli;
 
 import org.junit.Before;
 import org.junit.Test;
