+++ src/test/java/org/schemaspy/app/integrationtesting/DatabaseServiceIT.java
@@ -1,13 +1,15 @@
-package org.schemaspy.service;
+package org.schemaspy.app.integrationtesting;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.schemaspy.Config;
-import org.schemaspy.cli.CommandLineArguments;
+import org.schemaspy.app.cli.CommandLineArguments;
 import org.schemaspy.model.Database;
 import org.schemaspy.model.ProgressListener;
+import org.schemaspy.service.DatabaseService;
+import org.schemaspy.service.SqlService;
 import org.schemaspy.testing.H2MemoryRule;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
